+++ versions/v19/projects/sdcorejs-angular/components/form-generic/src/models/form-generic-component.model.ts
@@ -279,6 +279,7 @@ export interface SdFormGenericGroup {
   properties: {
     icon: string;
     color: Color;
+    collapsible?: boolean;
     hidden?: boolean;
     hiddenWhenExpression?: SdFormGenericExpression;
     visibleWhenExpression?: SdFormGenericExpression;